+++ pipeline/transformer/transformer_sierra/src/main/scala/uk/ac/wellcome/platform/transformer/sierra/transformers/SierraNotes.scala
@@ -16,6 +16,7 @@ object SierraNotes extends SierraDataTransformer with SierraQueryOps {
     "502" -> createNoteFromContents(DissertationNote),
     "504" -> createNoteFromContents(BibliographicalInformation),
     "505" -> createNoteFromContents(ContentsNote),
+    "506" -> createNoteFromContents(TermsOfUse),
     "508" -> createNoteFromContents(CreditsNote),
     "510" -> createNoteFromContents(PublicationsNote),
     "511" -> createNoteFromContents(CreditsNote),
